fix(test): use BigDecimal for product price in test

+++ backend/src/main/java/com/vantage/product/ui/dto/ProductResponse.java
@@ -1,6 +1,7 @@
 package com.vantage.product.ui.dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
-public record ProductResponse(UUID id, String name, String description, Double price) {
+public record ProductResponse(UUID id, String name, String description, BigDecimal price) {
 }
